Several tests for OrderService implemented. Code changes according to PR comments. Other minor test improvements. Product and ProductOrder Quantity property type modified to int and Price property modified to decimal according to previous PR

+++ EcommerceStore.Application/Models/InputModels/ProductInputModel.cs
@@ -9,9 +9,9 @@ namespace EcommerceStore.Application.Models.InputModels
         public string Name { get; set; }
         [Required]
         [Range(0.1, double.MaxValue, ErrorMessage = "Price must be positive floating-point numeric number")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         [Required]
-        public decimal Quantity { get; set; }
+        public int Quantity { get; set; }
         [Required]
         [StringLength(500, MinimumLength = 1, ErrorMessage = "Product's description must be between 1 and 500 characters")]
         public string Description { get; set; }
